不要なカスタムフックを削除
